% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{data}
\name{counts_compressed}
\alias{counts_compressed}
\title{A(nother) sample count matrix, just stored in bzip2 compressed format}
\description{
A(nother) sample count matrix, just stored in bzip2 compressed format
}
